pyproject.toml
[tool.poetry]
name = "engram-v3"
version = "0.1.0"
description = ""
authors = ["Arno Klein"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
pyyaml = "^6.0.2"
tqdm = "^4.67.1"
psutil = "^6.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
